Add option to simulate a CSP error when signing in
**Why:** we didn't have a way to trigger a CSP error due to a missing
redirect URI, which made it difficult to generate documentation on how
to troubleshoot and resolve said issues. This commit adds a checkbox to
the authentication options that, if selected, will redirect the user to
https://www.example.com after authentication, which should not be
configured as a redirect URI in the Dashboard. This will trigger a CSP
error on the redirect in Chrome and Chromium-based browsers.

Also includes some language updates around the authentication options
to remove references to IAL2.