Add a temporary biometric comparison flag
In order to test the biometric comparison experience on the IdP we want this app to make a request that requires a biometric comparison. The details of exactly what this will look like are still being ironed out.

This commit adds an option to require a biometric comparison which makes a request with a `biometric_comparison_required` value on the authorization URL. This is not how this is intended to work long term but it does let us test the functionality on the IdP.

At this time the behavior is not implemented on the IdP.