LG-12912 Rename `VTR_ENABLED` to `vtr_disabled` (#157)
The use of the VTR param has been authorized by security and is to be enabled in all environments except for prod. This commit renames the `VTR_ENABLED` config value that enabled the VTR param to `vtr_disabled` which does the opposite. This effectively changes the default behavior if no value is specified.

This commit downcases the name to match the style of the other configs.

We will need to correctly set `vtr_disabled=true` in prod and enable VTR support in the IdP in all environments except prod before merging this.