Add note to Biometric Comparison that it is disabled in prod
The biometric comparison feature has been enabled in all environments except for prod. We want to ensure that no partners see it enabled in int and try to use it in their apps which will not work when they are promoted in prod. This commit helps with that by adding a label to the "Biometric comparison" option on this sample app.